feat: add OpenSSL AES-128 backend of `Encrypter`
